# Improve Encoding.UTF8.GetMaxByte/CharCount perf