Mostly a very small perf improvement to `GetMaxByteCount` and `GetMaxCharCount`. It's now small enough to inline into callers when using the `Encoding.UTF8` accelerator.

Also addresses:
* Adds integer overflow protection to some unmanaged callers.
* Improves unit test coverage.
* Cleans up some XML usage.

|             Method |        Job | Toolchain |  input |     Mean |    Error |   StdDev | Ratio | RatioSD |
|------------------- |----------- |---------- |------- |---------:|---------:|---------:|------:|--------:|
| GetTranscodedLenth | Job-LLVNSR |  baseline | Hello! | 15.27 ns | 0.383 ns | 1.123 ns |  1.00 |    0.00 |
| GetTranscodedLenth | Job-DMFSKR |   compare | Hello! | 13.51 ns | 0.306 ns | 0.340 ns |  0.87 |    0.06 |

```cs
[Benchmark]
[Arguments("Hello!")]
public int GetTranscodedLenth(string input)
{
    const int MaxStackLength = 64;
    byte[] rentedArray = null;
    int maxByteCount = Encoding.UTF8.GetMaxByteCount(input.Length);
    Span<byte> scratch = ((uint)maxByteCount <= MaxStackLength)
        ? stackalloc byte[MaxStackLength]
        : (rentedArray = ArrayPool<byte>.Shared.Rent(maxByteCount));

    try
    {
        return Encoding.UTF8.GetBytes(input, scratch);
    }
    finally
    {
        if (rentedArray != null)
        {
            ArrayPool<byte>.Shared.Return(rentedArray);
        }
    }
}
```