@GrabYourPitchforks 2022-05-31
I updated the code comments in `UTF8Encoding` to more accurately describe why the `+1` is important. Latest iteration is a comment-only change; no functional changes from previous iteration.